feat: 集成mybatis,编写mapper供service层调用

+++ src/main/java/com/lobster/tobebraverman/service/UserService.java
@@ -5,12 +5,13 @@ import com.lobster.tobebraverman.entity.User;
 import java.util.List;
 
 public interface UserService {
-    User getUser(Long id);
+
+    User getUser(Integer id);
 
     List<User> getAll();
 
     User getUserByName(String name);
 
-    User getUser(Long id, String name);
+    User getUser(Integer id, String name);
 
 }
